Development: Extract web3 store
